Extract websocket context into a provider
