+++ src/Router/Authorized.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useEffect, useState } from 'react'
+import React from 'react'
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom'
 import { ApolloProvider } from '@apollo/react-hooks'
 import ApolloClient from 'apollo-boost'
@@ -9,17 +9,11 @@ import Invitations from 'Invitations'
 import Login from 'Login'
 import Room from 'Room'
 import Signup from 'Signup'
-
-import { Client, awaitWebsocket } from 'lib/WebsocketClient/client'
 import { API_HOST } from 'lib/constants'
 
-const socketClient = new Client({ debug: true })
-type WebsocketContext = InstanceType<typeof Client>
-export const WebsocketContext = createContext<WebsocketContext>(socketClient)
+import WebsocketContextProvider from './WebsocketContextProvider'
 
 const Router: React.FC<{ token: string }> = ({ token }) => {
-  const [socketConnected, setSocketConnected] = useState(false)
-
   const apolloClient = new ApolloClient({
     uri: `${API_HOST}/api/v1/graphql`,
     headers: {
@@ -27,20 +21,9 @@ const Router: React.FC<{ token: string }> = ({ token }) => {
     },
   })
 
-  useEffect(() => {
-    awaitWebsocket(token).then(websocket => {
-      socketClient.bind(websocket)
-      setSocketConnected(true)
-    })
-  }, [token])
-
-  if (!socketConnected) {
-    return <p>Loading...</p>
-  }
-
   return (
-    <ApolloProvider client={apolloClient}>
-      <WebsocketContext.Provider value={socketClient}>
+    <WebsocketContextProvider token={token}>
+      <ApolloProvider client={apolloClient}>
         <BrowserRouter>
           <Switch>
             <Route key="invitation" path="/invitation">
@@ -74,8 +57,8 @@ const Router: React.FC<{ token: string }> = ({ token }) => {
             />
           </Switch>
         </BrowserRouter>
-      </WebsocketContext.Provider>
-    </ApolloProvider>
+      </ApolloProvider>
+    </WebsocketContextProvider>
   )
 }
